+++ lib/sucker_punch/exceptions.rb
@@ -2,5 +2,4 @@ module SuckerPunch
   class Error < StandardError; end
   class MissingQueueName < Error; end
   class MissingWorkerName < Error; end
-
 end